02.21
일반게시판 조회

+++ JSP_Project/WebContent/views/board/boardDetailView.jsp
@@ -1,5 +1,12 @@
+<%@ page import="com.kh.borard.model.vo.*" %>
+
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
+<%
+	Board b = (Board)request.getAttribute("b");
+	Attachment at =(Attachment)request.getAttribute("at");
+%>
+
 <!DOCTYPE html>
 <html>
 <head>
@@ -22,20 +29,20 @@
       <table id="detail-area" align="center" border="1">
          <tr>
             <th width="70">카테고리</th>
-            <th width="70">운동</th>
+            <th width="70"><%= b.getCategory() %></th>
             <th width="70">제목</th>
-            <td width="350">제목부분입니다.</td>
+            <td width="350"><%= b.getBoardTitle() %></td>
          </tr>
          <tr>
             <th>작성자</th>
-            <td>user01</td>
+            <td><%= b.getBoardWriter() %></td>
             <th>작성일</th>
-            <td>2023-02-20</td>
+            <td><%= b.getCreateDate() %></td>
          </tr>
          <tr>
             <th>내용</th>
             <td colspan="3">
-               <p style="height:200px">글내용글내용글내용</p>
+               <p style="height:200px"><%= b.getBoardContent() %></p>
             </td>
          </tr>
          <tr>
@@ -43,7 +50,15 @@
             <td colspan="3">
                <!-- 첨부파일이 없을경우 : 첨부파일이 없습니다로 표시
                첨부파일이 있을경우 :  -->
-               <a download="파일원본명" href="해당파일이존재하는경로">파일원본명</a>
+               
+               <% if(at == null){%> <!-- at에 저장된 첨부파일이 없는 경우 -->
+               		첨부파일이 없습니다.
+               
+               <%}else{ %>
+               		<!--  href = '/jspproject/resources/board_upfiles/202xxxxxx.jpg -->
+	               <a download="<%=at.getOriginName() %>" href="<%=contextPath%>/<%=at.getFilePath()+at.getChangeName()%>"><%= at.getOriginName() %></a> <!-- filePath : 파일의 경로가 저장되어 있음 -->
+               <%} %>
+            
             </td>
          </tr>
       </table>
@@ -54,10 +69,10 @@
          <a href="<%=contextPath %>/list.bo?currentPage=1" class="btn btn-secondary btn-sm">목록가기</a>
       
          <!-- 로그인한 사용자가 해당 게시글의 작성자인 경우 -->
-         <%--  <% if(loginUser != null && loginUser.getUserId().equals(b.getBoardWriter() ) {
-            <a href="<%= contextPath %>/updateForm.bo?bno=<%=b.getBoardNo() %>" class="btn btn-warning btn-sm">수정하기</a>
+         <% if(loginUser != null && loginUser.getUserId().equals(b.getBoardWriter() )) { %>
+            <a href="<%= contextPath %>/update.bo?bno=<%=b.getBoardNo() %>" class="btn btn-warning btn-sm">수정하기</a>
             <a href="<%= contextPath %>/delete.bo?bno=<%=b.getBoardNo() %>" class="btn btn-danger btn-sm">삭제하기</a>
-         <% } %> --%>
+         <% } %> 
       </div>
       
       <br>
